+++ src/app/addItems/style.sass
@@ -1,2 +1,8 @@
 h1, label, button
     font-family: 'god-of-war'
+
+#fondPage
+    background-image: url('/seikret-mount.jpg')
+
+#addItem
+    background-color: rgba(217, 131, 36, 0.50)
